Added penalty for gutters

+++ code/entity/ScoreReport.java
@@ -2,7 +2,7 @@ package entity;
 
 /**
  * 
- * SMTP implementation based on code by Réal Gagnon mailto:real@rgagnon.com
+ * SMTP implementation based on code by Rï¿½al Gagnon mailto:real@rgagnon.com
  * 
  */
 
@@ -22,7 +22,7 @@ public class ScoreReport {
 
 	private String content;
 	
-	public ScoreReport(Bowler bowler, int[] scores, int games ) {
+	public ScoreReport(Bowler bowler, float[] scores, int games ) {
 		String nick = bowler.getNick();
 		String full = bowler.getFullName();
 		Vector v = null;
